+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # correctBmatrixOverflow changelog
 
 ## [unreleased]
+- added ability to apply a transformation before saving b-vectors
+- removed default DICOM output basename
+- increased output precision
 
 ## [v0.1.0]
 Initial public release